+++ types/src/main/java/com/cobber/fta/PluginDefinition.java
@@ -19,7 +19,6 @@ import java.io.BufferedReader;
 import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Locale;
@@ -53,8 +52,8 @@ public class PluginDefinition {
 	public String description;
 	/** Type of the plugin - can be 'java', 'list', or 'regex' */
 	public String pluginType;
-	/** Plugin-specific options - format key1=value1, key2=value2, ... */
-	public String pluginOptions;
+	/** Plugin-specific options as a JSON object */
+	public Map<String, Object> pluginOptions;
 	/** Signature (structure) - the MD5 Hash of the Semantic Type name and the Base Type. */
 	public String signature;
 	/** locales this plugin applies to - empty set, implies all locales.  Can use just language instead of tag, e.g. "en" rather than "en_US". */
@@ -89,8 +88,6 @@ public class PluginDefinition {
 	public int minSamples = -1;
 	/** Need to see both the minimum and maximum values to declare success. */
 	public boolean minMaxPresent;
-
-	private volatile Map<String, String> options = null;
 
 	public enum Precedence {
 	    BUILTIN,
@@ -287,21 +284,7 @@ public class PluginDefinition {
 		return ret.toString();
 	}
 
-	public Map<String, String> getOptions() {
-		if (options == null)
-			synchronized(this) {
-				if (options == null) {
-					options = new HashMap<>();
-					if (pluginOptions != null) {
-						final String[] entries = pluginOptions.split("\\s*,\\s*");
-						for (final String entry : entries) {
-							final int separator = entry.indexOf('=');
-							options.put(entry.substring(0,separator), entry.substring(separator + 1));
-						}
-					}
-				}
-			}
-
-		return options;
+	public Map<String, Object> getOptions() {
+		return pluginOptions;
 	}
 }